adk/google-adk/requirements.txt
# Google Cloud & AI
google-cloud-aiplatform>=1.38.0,<2.0.0
google-generativeai>=0.8.0,<1.0.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database (optional - using backend API instead)
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# HTTP Client (for backend communication)
requests>=2.31.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
python-multipart>=0.0.6

# Date & Time
python-dateutil>=2.8.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# WebSocket Client
python-socketio[client]>=5.11.0